266: AbstractTest in Java
Task-Url: http://github.com/LorenzoBettini/edelta/issues/266
